refactor: cull some warnings

+++ src/tier1/utlstring.cpp
@@ -478,7 +478,7 @@ void CUtlString::TrimLeft( char cTarget )
 	// We have some whitespace to remove
 	if ( nIndex > 0 )
 	{
-		memcpy( m_pString, &m_pString[nIndex], Length() - nIndex );
+		memcpy( m_pString, &m_pString[nIndex], ( size_t )( Length() - nIndex ) );
 		SetLength( Length() - nIndex );
 	}
 }
@@ -515,7 +515,7 @@ void CUtlString::TrimLeft( const char *szTargets )
 	// We have some whitespace to remove
 	if ( i > 0 )
 	{
-		memcpy( m_pString, &m_pString[i], Length() - i );
+		memcpy( m_pString, &m_pString[i], ( size_t )( Length() - i ) );
 		SetLength( Length() - i );
 	}
 }
